Exercise_11 Version 2

+++ vera.galstyan/repo/Lesson_3/exercise_11/exercise_11.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python3
 
 
-width = float(input("Enter a width of farmer's field in feet: "))
+width =int(input("Enter a width of farmer's field in feet: "))
 
 
-length = float(input("Enter a length of farmer's field in feet: "))
+length = int(input("Enter a length of farmer's field in feet: "))
 
 
 square_feet_in_acre = 43560
 
 
-print("Total area of the field is", (width*length) /square_feet_in_acre, "acre") 
+print ("Total area of the field is", (width*length) /square_feet_in_acre, "acre")
